pyproject.toml
[tool.poetry]
name = "spkrepo"
version = "0.2.2"
description = "Synology Package Repository"
license = "MIT"
authors = ["Antoine Bertin <diaoulael@gmail.com>"]
readme = "README.md"
repository = "https://github.com/SynoCommunity/spkrepo"
keywords = ["synology", "package", "repository"]
homepage = "https://synocommunity.com"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Framework :: Flask",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
    "Topic :: System :: Archiving :: Packaging",
]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.2.0"
flask-sqlalchemy = "^2.4.1"
passlib = "^1.7.2"
flask-babelex = "^0.9.4"
flask-wtf = "^1.0.0"
wtforms = "^3.0.1"
flask-mail = "^0.9.1"
configparser = "^5.0.0"
email_validator = "^1.1.1"
flask-principal = "^0.4.0"
flask-admin = "^1.5.6"
sqlalchemy = "^1.3.17"
pillow = "^8.3.2"
flask-restful = "^0.3.8"
flask-login = "^0.6.0"
flask-caching = "^1.8.0"
redis = "^4.1.0"
python-gnupg = "^0.4.6"
requests = "^2.23.0"
click = "^8.0.0"
flask-migrate = "^3.0.0"
alembic = "^1.4.2"
text-unidecode = "^1.3"
ipaddress = "^1.0.23"
flask-debugtoolbar = "^0.11.0"
bcrypt = "^3.1.7"
Flask-CLI = "^0.4.0"
Flask-Security-Too = {extras = ["common", "fsqla"], version = "^5.0.2"}
Flask = "^2.2.2"

[tool.poetry.dev-dependencies]
sphinx = "^3.0.3"
sphinx-rtd-theme = "^0.4.3"
flask-testing = "^0.8.0"
factory-boy = "^2.12.0"
faker = "^4.1.0"
lxml = "^4.5.1"
mock = "^4.0.2"
pre-commit = "^2.4.0"
pytest = "^6.2.5"

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
known_first_party = "spkrepo"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
